+++ bridge-forwarder/src/mock.rs
@@ -3,10 +3,12 @@
 
 #![cfg(test)]
 
+use cumulus_primitives_core::ParaId;
 use std::marker::PhantomData;
 use std::result;
 
 use frame_support::dispatch::DispatchResult;
+use frame_support::pallet_prelude::Get;
 use frame_support::{
 	construct_runtime, parameter_types,
 	traits::{AsEnsureOriginWithArg, ConstU32},
@@ -18,7 +20,7 @@ use sp_runtime::testing::H256;
 use sp_runtime::traits::{BlakeTwo256, IdentityLookup};
 use sp_runtime::{AccountId32, BuildStorage};
 use xcm::latest::{BodyId, Junction, MultiAsset, MultiLocation, NetworkId};
-use xcm::prelude::{Concrete, Fungible, GeneralKey, Parachain, X3};
+use xcm::prelude::{Concrete, Fungible, GeneralKey, Parachain, X1, X3};
 use xcm::v3::Weight;
 use xcm_builder::{
 	AccountId32Aliases, CurrencyAdapter, FungiblesAdapter, IsConcrete, NoChecking, ParentIsPreset,
@@ -26,7 +28,7 @@ use xcm_builder::{
 };
 use xcm_executor::traits::{Error as ExecutionError, MatchesFungibles};
 
-use sygma_traits::{Bridge, TransactorForwarder};
+use sygma_traits::{AssetTypeIdentifier, Bridge, TransactorForwarder};
 
 use crate as sygma_bridge_forwarder;
 
@@ -256,6 +258,25 @@ impl MatchesFungibles<AssetId, Balance> for SimpleForeignAssetConverter {
 				}
 			},
 			_ => Err(ExecutionError::AssetNotHandled),
+		}
+	}
+}
+
+/// NativeAssetTypeIdentifier impl AssetTypeIdentifier for XCMAssetTransactor
+/// This impl is only for local mock purpose, the integrated parachain might have their own version
+pub struct NativeAssetTypeIdentifier<T>(PhantomData<T>);
+impl<T: Get<ParaId>> AssetTypeIdentifier for NativeAssetTypeIdentifier<T> {
+	/// check if the given MultiAsset is a native asset
+	fn is_native_asset(asset: &MultiAsset) -> bool {
+		// currently there are two multilocations are considered as native asset:
+		// 1. integrated parachain native asset(MultiLocation::here())
+		// 2. other parachain native asset(MultiLocation::new(1, X1(Parachain(T::get().into()))))
+		let native_locations =
+			[MultiLocation::here(), MultiLocation::new(1, X1(Parachain(T::get().into())))];
+
+		match (&asset.id, &asset.fun) {
+			(Concrete(ref id), Fungible(_)) => native_locations.contains(id),
+			_ => false,
 		}
 	}
 }
